Convert produce to return unique_ptr, fix mkesprod
In preparation for concurrent IOVs, convert ESProducer
produce methods to return unique_ptr.

This affects modules used in Framework unit tests.
Also modifies the Skeletons script that will generate
a template for a new ESProducer.

Note that Skeletons script had been broken by the
changes to calls to the Python print function which
were preparing for the migration to Python 3. This
fixes that problem also.